Move tests to complex.jl
