+++ test/complex.jl
@@ -1065,3 +1065,79 @@ end
 
     @test @inferred(2.0^(3.0+0im)) === @inferred((2.0+0im)^(3.0+0im)) === @inferred((2.0+0im)^3.0) === 8.0+0.0im
 end
+
+
+# Tests for f(a+ib) for a, b not AbstractFloats
+# 25292, 25793
+
+## Simple float wrapper type:
+
+struct FloatWrapper <: Real
+    x::Float64
+end
+
+import Base: +, -, *, /, ^, <, <=, ==, ≈,
+    sqrt, sin, cos, tan, sinh, cosh, tanh,
+    exp, expm1, exp2, exp10,
+    log, log1p, log2, log10,
+    asin, acos, atan, atan2,
+    asinh, acosh, atanh,
+    convert, isfinite, float, promote_rule
+
+for op in (:+, :-, :*, :/, :^, :atan2)
+    @eval $op(x::FloatWrapper, y::FloatWrapper) = FloatWrapper($op(x.x, y.x))
+end
+
+for op in (:-, :sqrt, :sin, :cos, :tan,
+        :sinh, :cosh, :tanh,
+        :exp, :exp2, :exp10, :expm1,
+        :log, :log1p, :log2, :log10,
+        :asin, :acos, :atan,
+        :asinh, :acosh, :atanh)
+
+    @eval $op(x::FloatWrapper) = FloatWrapper($op(x.x))
+end
+
+for op in (:isfinite,)
+    @eval $op(x::FloatWrapper) = $op(x.x)
+end
+
+convert(::Type{FloatWrapper}, x::Int) = FloatWrapper(float(x))
+promote_rule(::Type{FloatWrapper}, ::Type{Int}) = FloatWrapper
+
+convert(::Type{FloatWrapper}, x::Float64) = FloatWrapper(x)
+promote_rule(::Type{FloatWrapper}, ::Type{Float64}) = FloatWrapper
+
+
+float(x::FloatWrapper) = x
+
+<(x::FloatWrapper, y::FloatWrapper) = (x.x < y.x)
+==(x::FloatWrapper, y::FloatWrapper) = (x.x == y.x)
+<=(x::FloatWrapper, y::FloatWrapper) = (x.x <= y.x)
+≈(x::FloatWrapper, y::FloatWrapper) = (x.x ≈ y.x)
+
+Base.rtoldefault(::Type{FloatWrapper})  = Base.rtoldefault(Float64)
+
+@testset "f(a+ib) for a, b that are not AbstractFloats; #25292" begin
+
+    x, y = 3.1, 4.1
+    z = x + im*y
+    wrapped_z = FloatWrapper(x) + im * FloatWrapper(y)
+
+    @test sincos(x) == (sin(x), cos(x))
+
+    for f in (inv,
+              exp, expm1,
+              sin, cos, tan,
+              sinh, cosh, tanh,
+              log, log2, log10, log1p,
+              asin, acos,
+              asinh, acosh)
+
+        z2 = f(wrapped_z)
+        z3 = f(z)
+
+        @test isa(z2, Complex)
+        @test z2 ≈ FloatWrapper(z3.re) + im*FloatWrapper(z3.im)
+    end
+end
